Use long doubles to store exp vectors used to compute markage.
This mitigates the issues with exp(t/tau) and exp(-t/tau) terms
over/underflowing when tau is too small with respect to the spike
times: on my system, if tau=1ms this happened for spike times larger
than a few hundred ms. Using long doubles, I can go up to about 10s.

Overall, there's still work to be done to endure that when the
exponential do blow up we raise an exception at the Python level, as
opposed to what's happening now (all distances being silently set to
zero).

+++ include/Van_Rossum_Multiunit.hpp
@@ -15,14 +15,14 @@ void d_exp_markage(double **d_matrix,vector<vector<vector<double> > > & trains, 
 void d_exp_markage_rect(double **d_matrix,vector<vector<vector<double> > > & trains1,vector<vector<vector<double> > > & trains2, double tau,double c);
 
 double big_r(vector<double> & u,double tau);
-double big_r_with_exp(vector<double> & u,vector<double> & exp_p,vector<double> & exp_m,double tau);
+double big_r_with_exp(vector<double> & u,vector<long double> & exp_p,vector<long double> & exp_m,double tau);
 double big_r_with_exp_markage(vector<double> & fs);
 
 double big_r(vector<double> & u_a,vector<double> & u_b,double tau);
-double big_r_with_exp(vector<double> & u_a,vector<double> & exp_p_a,vector<double> & exp_m_a,vector<double> & u_b,vector<double> & exp_p_b,vector<double> & exp_m_b,double tau);
-double big_r_with_exp_markage(vector<double> & train_a, vector<double> & f_a,vector<double> & e_pos_a,vector<double> & e_neg_a,vector<double> & train_b, vector<double> & f_b,vector<double> & e_pos_b,vector<double> & e_neg_b);
+double big_r_with_exp(vector<double> & u_a,vector<long double> & exp_p_a,vector<long double> & exp_m_a,vector<double> & u_b,vector<long double> & exp_p_b,vector<long double> & exp_m_b,double tau);
+double big_r_with_exp_markage(vector<double> & train_a, vector<double> & f_a,vector<long double> & e_pos_a,vector<long double> & e_neg_a,vector<double> & train_b, vector<double> & f_b,vector<long double> & e_pos_b,vector<long double> & e_neg_b);
 
-void expage(vector<double> & e_pos, vector<double> & e_neg,vector<double> & train,double tau);
-void markage(vector<double> & f, vector<double> & e_pos, vector<double> & e_neg, vector<double> & train,double tau);
+void expage(vector<long double> & e_pos, vector<long double> & e_neg,vector<double> & train,double tau);
+void markage(vector<double> & f, vector<long double> & e_pos, vector<long double> & e_neg, vector<double> & train,double tau);
 
 #endif
